refactor: Service 역할 분리
